converted integer positions to floats, solved chaining issues

+++ src/Animatable.h
@@ -6,27 +6,27 @@
 #include "RenderUtils.h"
 using namespace std;
 struct Position {
-	int x;
-	int y;
+	float x;
+	float y;
 };
 class AnimationElement {
 	AnimationElement* next;
-	int x;
-	int y;
-	int radius;
+	float x;
+	float y;
+	float radius;
 	SDL_Color colour;
-	Position MoveElement(int prevX, int prevY, int prevRadius);
+	Position MoveElement(float prevX, float prevY, int prevRadius);
 public:
 	AnimationElement(int radius, SDL_Color colour, int positionOffset);
 	void SetNext(AnimationElement* nextToSet);
-	void UpdatePosition(int prevX, int prevY, int prevRadius);
+	void UpdatePosition(float prevX, float prevY, int prevRadius);
 	void Render(SDL_Renderer* renderer);
 };
 class Animatable {
 	int chainLength;
 	AnimationElement* first;
-	int x;
-	int y;
+	float x;
+	float y;
 public:
 	Animatable(vector<int> radii, vector<SDL_Color> colours);
 	void Update();
